+++ libraries/import/ods.php
@@ -124,17 +124,17 @@ foreach ($sheets as $sheet) {
                     for ($k = 0; $k < $num_iterations; $k++) {
                         if (! $col_names_in_first_row) {
                             if ($_REQUEST['ods_recognize_percentages'] && !strcmp('percentage', $cell_attrs['value-type'])) {
-                                $tempRow[] = (double)$cell_attrs['value'];
+                                $tempRow[] = (float)$cell_attrs['value'];
                             } elseif ($_REQUEST['ods_recognize_currency'] && !strcmp('currency', $cell_attrs['value-type'])) {
-                                $tempRow[] = (double)$cell_attrs['value'];
+                                $tempRow[] = (float)$cell_attrs['value'];
                             } else {
                                 $tempRow[] = (string)$text;
                             }
                         } else {
                             if ($_REQUEST['ods_recognize_percentages'] && !strcmp('percentage', $cell_attrs['value-type'])) {
-                                $col_names[] = (double)$cell_attrs['value'];
+                                $col_names[] = (float)$cell_attrs['value'];
                             } else if ($_REQUEST['ods_recognize_currency'] && !strcmp('currency', $cell_attrs['value-type'])) {
-                                $col_names[] = (double)$cell_attrs['value'];
+                                $col_names[] = (float)$cell_attrs['value'];
                             } else {
                                 $col_names[] = (string)$text;
                             }
